\name{runshinySQLite}
\alias{runshinySQLite}
\title{Hello, World!}
\usage{
runshinySQLite()
}
\description{
Runs the demo shiny SQLite app. You can use this demo app to do your SQLite databases since it allows you to create your new and custom databases.
}
\examples{
Run the following function
runshinySQLite()
}
