\name{runshinySQLite}
\alias{runshinySQLite}
\title{Run and use shiny SQLite demo app}
\usage{
runshinySQLite()
}
\description{
Runs the demo shiny SQLite app. You can use this demo app to do your SQLite databases since it allows you to create your new and custom databases.
}
\examples{
Run 
runshinySQLite()
}
